requirements.txt
click>=8.1.0
selenium>=4.15.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.2
keyring>=24.0.0
rich>=13.0.0
webdriver-manager>=4.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pytest-cov>=4.1.0